+++ CHANGES.rst
@@ -35,6 +35,8 @@
   unmaintained, and usage is shifting from Webassets to Webpack
 * ``for_tsquery`` has been removed as PostgreSQL>=12 has native functions
 * ``render_with`` no longer offers a shorthand for JSONP responses
+* ``coaster.sqlalchemy.ModelBase`` now replaces Flask-SQLAlchemy's db.Model
+  with full support for type hinting
 
 0.6.1 - 2021-01-06
 ------------------